+++ PSql/_Commands/ExpandSqlCmdDirectivesCommand.cs
@@ -7,7 +7,7 @@ namespace PSql;
 
 [Cmdlet(VerbsData.Expand, "SqlCmdDirectives")]
 [OutputType(typeof(string[]))]
-public class ExpandSqlCmdDirectivesCommand : Cmdlet
+public class ExpandSqlCmdDirectivesCommand : PSCmdlet
 {
     // -Sql
     [Parameter(Position = 0, Mandatory = true, ValueFromPipeline = true)]